+++ ios/MullvadVPN/Containers/Root/RootConfiguration.swift
@@ -1,0 +1,15 @@
+//
+//  RootConfiguration.swift
+//  MullvadVPN
+//
+//  Created by Jon Petersson on 2023-04-19.
+//  Copyright © 2023 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+
+struct RootConfigration {
+    var deviceName: String?
+    var expiry: Date?
+    var showsAccountButton: Bool
+}