+++ ios/MullvadVPN/Presentation controllers/SecondaryContextPresentationController.swift
@@ -21,7 +21,7 @@ class SecondaryContextPresentationController: FormsheetPresentationController {
         if let containerView = containerView,
            let rootContainer = presentingViewController as? RootContainerViewController
         {
-            rootContainer.addSettingsButtonToPresentationContainer(containerView)
+            rootContainer.addTrailingButtonsToPresentationContainer(containerView)
         }
     }
 
@@ -29,7 +29,7 @@ class SecondaryContextPresentationController: FormsheetPresentationController {
         super.dismissalTransitionDidEnd(completed)
 
         if let rootContainer = presentingViewController as? RootContainerViewController, completed {
-            rootContainer.removeSettingsButtonFromPresentationContainer()
+            rootContainer.removeTrailingButtonsFromPresentationContainer()
         }
     }